Rename: graph -> interface/graphql
