+++ backend/cmd/server/main.go
@@ -19,11 +19,11 @@ import (
 	"github.com/oinume/lekcije/backend/di"
 	"github.com/oinume/lekcije/backend/event_logger"
 	"github.com/oinume/lekcije/backend/gcp"
-	"github.com/oinume/lekcije/backend/graph/generated"
-	"github.com/oinume/lekcije/backend/graph/resolver"
 	"github.com/oinume/lekcije/backend/infrastructure/ga_measurement"
 	"github.com/oinume/lekcije/backend/infrastructure/mysql"
 	interfaces "github.com/oinume/lekcije/backend/interface"
+	"github.com/oinume/lekcije/backend/interface/graphql/generated"
+	"github.com/oinume/lekcije/backend/interface/graphql/resolver"
 	interfaces_http "github.com/oinume/lekcije/backend/interface/http"
 	"github.com/oinume/lekcije/backend/logger"
 	"github.com/oinume/lekcije/backend/model"
