+++ backend/gqlgen.yml
@@ -1,26 +1,26 @@
 # Where are all the schema files located? globs are supported eg  src/**/*.graphqls
 schema:
-  - graph/schema/*.graphqls
+  - interface/graphql/schema/*.graphqls
 
 # Where should the generated server code go?
 exec:
-  filename: graph/generated/generated.go
+  filename: interface/graphql/generated/generated.go
   package: generated
 
 # Uncomment to enable federation
 # federation:
-#   filename: graph/generated/federation.go
+#   filename: interface/graphql/generated/federation.go
 #   package: generated
 
 # Where should any generated models go?
 model:
-  filename: graph/model/models_gen.go
+  filename: interface/graphql/model/models_gen.go
   package: model
 
 # Where should the resolver implementations go?
 resolver:
   layout: follow-schema
-  dir: graph/resolver
+  dir: interface/graphql/resolver
   package: resolver
   filename_template: "{name}.resolvers.go"
 
@@ -36,7 +36,7 @@ resolver:
 # gqlgen will search for any type names in the schema in these go packages
 # if they match it will use them, otherwise it will generate them.
 autobind:
-  - "github.com/oinume/lekcije/backend/graph/model"
+  - "github.com/oinume/lekcije/backend/interface/graphql/model"
 
 # This section declares type mapping between the GraphQL and go type systems
 #
